+++ packages/errors/index.js
@@ -54,9 +54,9 @@ if (missing.length > 0) {
 // and also updates the names of the utility functions.
 const {
   bare,
-  details: redacted,
+  details,
   error,
-  Fail: throwRedacted,
+  Fail,
   makeAssert: _omittedMakeAssert,
   makeError,
   note,
@@ -81,19 +81,21 @@ const bestMakeError = makeError || error;
 export {
   // assertions
   assert,
-  // related utilities that aren't assertions
+  // non-assertion utilities that appear as properties of `assert`
   bareOrQuote as bare,
   bestMakeError as makeError,
+  details,
   note,
   quote,
-  redacted,
-  throwRedacted,
-  // conventional abbreviations and aliases
+  Fail,
+  // conventional abbreviations
   bareOrQuote as b,
+  details as X,
   quote as q,
-  redacted as X,
-  throwRedacted as Fail,
+  // other aliases
   note as annotateError,
+  details as redacted,
+  Fail as throwRedacted,
 };
 
 /**
@@ -116,7 +118,7 @@ export {
  * @returns {T}
  */
 export const hideAndHardenFunction = func => {
-  typeof func === 'function' || throwRedacted`${func} must be a function`;
+  typeof func === 'function' || Fail`${func} must be a function`;
   const { name } = func;
   defineProperty(func, 'name', {
     // Use `String` in case `name` is a symbol.